requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.108.0
urllib3==2.1.0
tqdm==4.66.1